pyproject.toml
[tool.poetry]
name = "monnani-back"
version = "0.1.0"
description = ""
authors = ["remeo08 <rmeorux5870@gmail.com>"]
readme = "README.md"
packages = [{include = "monnani_back"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
